feat: permissions mapping  - "can" rules in routes for managing features and comments  - hide feature managing rights for users without manage_feature permission  - hide writing comment for users without manage_comment permission  - create helper for setting permission  - add rule on destroy comment function

+++ app/Http/Controllers/CommentController.php
@@ -28,6 +28,11 @@ class CommentController extends Controller
 
    public function destroy(Comment $comment)
   {
+    // TODO: create a gate/policy to replace this
+    if ($comment->user_id !== Auth::id()) {
+      abort(403, 'Unauthorized action.');
+    }
+
     $featureId = $comment->feature_id;
     $comment->delete();
